UCS is still adding duplicate nodes to Node[] PathNodes
when the test case gets larger in size. In one of the test case, the
founded goal node has a path of over 10,000+ nodes, while the longest
path we can expect is 500.

Need to update the algo, either remove the pass by reference part, or
implement with the parent style