printing some duplicate jumps
currently jumped 32 times, there are still some duplicates in the jump
stack.
jumped from 1,0 to 3,2
jumped from 3,2 to 5,4
jumped from 3,2 to 5,0
jumped from 2,0 to 4,2
jumped from 4,2 to 4,4
jumped from 3,0 to 5,2
jumped from 5,2 to 3,4
jumped from 3,0 to 5,0
jumped from 3,0 to 3,2
jumped from 3,2 to 5,4
jumped from 3,2 to 5,0
jumped from 4,0 to 4,2
jumped from 4,2 to 4,4
jumped from 1,1 to 3,3
jumped from 3,3 to 5,3
jumped from 3,1 to 5,1
jumped from 0,2 to 2,4
jumped from 1,2 to 3,4
jumped from 1,2 to 3,2
jumped from 3,2 to 5,4
jumped from 3,2 to 5,0
jumped from 2,2 to 2,4
jumped from 0,3 to 2,5
jumped from 0,3 to 0,5
jumped from 1,3 to 3,3
jumped from 3,3 to 5,3
jumped from 1,3 to 1,5
jumped from 2,3 to 0,5
jumped from 0,4 to 2,4
jumped from 1,4 to 3,2
jumped from 3,2 to 5,4
jumped from 3,2 to 5,0
minion 31 completed!
jumped 32 times
hw2 terminated